chore: refactor `Mailer`
