+++ apps/meteor/client/views/admin/mailer/Mailer.stories.tsx
@@ -1,15 +1,15 @@
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 import React from 'react';
 
-import { Mailer } from './Mailer';
+import MailerPage from './MailerPage';
 
 export default {
 	title: 'Admin/Mailer/Mailer',
-	component: Mailer,
+	component: MailerPage,
 	parameters: {
 		layout: 'fullscreen',
 	},
-} as ComponentMeta<typeof Mailer>;
+} as ComponentMeta<typeof MailerPage>;
 
-export const Default: ComponentStory<typeof Mailer> = (args) => <Mailer {...args} />;
+export const Default: ComponentStory<typeof MailerPage> = () => <MailerPage />;
 Default.storyName = 'Mailer';
